+++ test/javascript/events/pointer_event_test.mjs
@@ -21,8 +21,8 @@ describe("PointerEvent", () => {
       assert.deepStrictEqual(
         result,
         Type.map([
-          [Type.atom("page_x"), Type.integer(1)],
-          [Type.atom("page_y"), Type.integer(2)],
+          [Type.atom("page_x"), Type.float(1)],
+          [Type.atom("page_y"), Type.float(2)],
           [Type.atom("pointer_type"), Type.atom("mouse")],
         ]),
       );
@@ -35,8 +35,8 @@ describe("PointerEvent", () => {
       assert.deepStrictEqual(
         result,
         Type.map([
-          [Type.atom("page_x"), Type.integer(1)],
-          [Type.atom("page_y"), Type.integer(2)],
+          [Type.atom("page_x"), Type.float(1)],
+          [Type.atom("page_y"), Type.float(2)],
           [Type.atom("pointer_type"), Type.nil()],
         ]),
       );
